Change from oiejq to sio2jail (#254)
* Change to `sio2jail`

* Tests

* Fix for init command

* Fix various files

* Fix linux workflows

* Revert "Fix linux workflows"

This reverts commit 22e11bdc5543e24070cdb7591a1b5b3b32f6161c.

* Some debug

* Fix sio2jail executor

+++ src/sinol_make/commands/init/__init__.py
@@ -38,7 +38,8 @@ class Command(BaseCommand):
         if ret.returncode != 0:
             util.exit_with_error("Could not access repository. Please try again.")
         path = os.path.join(tmp_dir, package_dir)
-        shutil.rmtree(os.path.join(path, '.git'))
+        if os.path.exists(os.path.join(path, '.git')):
+            shutil.rmtree(os.path.join(path, '.git'))
         return path
 
     def move_folder(self):
